+++ app/component/userinfo/type.ts
@@ -1,0 +1,12 @@
+/**
+ * @author orangeboyChen
+ * @version 1.0
+ * @date 2025/2/8 01:31
+ */
+
+enum AvatarListLinkType {
+	HOME = 'HOME',
+	LOGOUT = 'LOGOUT',
+}
+
+export { AvatarListLinkType };
