+++ app/course-grade-stat/page.tsx
@@ -1,13 +1,16 @@
-'use client';
-
 import { Search } from '@/app/course-grade-stat/search';
 import Header from '@/app/course-grade-stat/header';
+import { Metadata } from 'next';
+
+export const metadata: Metadata = {
+	title: '给分搜索 | Ham',
+};
 
 const Home = () => {
 	return (
 		<div
 			className={
-				'grid grid-rows-[60px_0.8fr_0.2fr_60px] items-center justify-items-center min-h-screen px-4 sm:px-20'
+				'grid grid-rows-[104px_0.8fr_0.2fr_60px] items-center justify-items-center min-h-screen'
 			}
 		>
 			<Header />
