+++ app/course-grade-stat/search/page.tsx
@@ -1,45 +1,25 @@
-'use client';
+import SearchResultClient from '@/app/course-grade-stat/search/page.client';
+import { Metadata } from 'next';
 
 /**
  * @author orangeboyChen
  * @version 1.0
  * @date 2025/1/26 15:40
  */
-import { Header } from '@/app/course-grade-stat/search/header';
-import Body from '@/app/course-grade-stat/search/body';
-import { useEffect, useState } from 'react';
-import {
-	JsCourseGradeStatisticsItemOrder,
-	JsCourseScoreItem,
-	JsCourseScoreItemType,
-	JsScoreService,
-} from '@/wasm/pkg';
-import { useSearchParams } from 'next/navigation';
-import useRequest from '@/app/common/request';
+
+export async function generateMetadata({
+	searchParams,
+}: {
+	searchParams: Promise<{ [key: string]: string | string[] | undefined }>;
+}): Promise<Metadata> {
+	const params = (await searchParams) as never;
+	return {
+		title: params['keyword'] + ' | Ham',
+	};
+}
 
 const SearchResult = () => {
-	const query = useSearchParams();
-	const keyword = query.get('keyword') ?? '';
-	const [result, setResult] = useState<JsCourseScoreItem[]>([]);
-	const { request } = useRequest();
-	useEffect(() => {
-		request({
-			call: () =>
-				JsScoreService.getScoreStat(
-					JsCourseScoreItemType.courseName(),
-					keyword,
-					JsCourseGradeStatisticsItemOrder.scoreDesc()
-				),
-		}).then((r) => setResult(r));
-	}, [keyword]);
-
-	return (
-		<div className={'min-h-screen h-full bg-black/5'}>
-			<div className={'h-[8px] bg-white'} />
-			<Header queryKeyword={keyword} />
-			<Body result={result} />
-		</div>
-	);
+	return <SearchResultClient />;
 };
 
 export default SearchResult;
