+++ app/course-grade-stat/search/search.ts
@@ -1,0 +1,3 @@
+const useSearch = () => {};
+
+export default useSearch;
