+++ app/service/auth/index.ts
@@ -1,0 +1,3 @@
+import useAuth from '@/app/service/auth/auth';
+
+export { useAuth };
