+++ wasm/pkg/ham_web_wasm_bg.wasm.d.ts
@@ -41,14 +41,14 @@ export const __wbg_set_jssearchscorehititem_hit: (a: number, b: number, c: numbe
 export const __wbg_get_jssearchscorehititem_value: (a: number) => [number, number];
 export const __wbg_get_jssearchscorehititem_hit_type: (a: number) => [number, number];
 export const __wbg_get_jssearchscorehititem_hit: (a: number) => [number, number];
-export const start: () => void;
+export const __wbg_jspasskeyservice_free: (a: number, b: number) => void;
+export const jspasskeyservice_get_credentials_request_options: () => any;
 export const __wbg_jsrequesterror_free: (a: number, b: number) => void;
 export const __wbg_get_jsrequesterror_code: (a: number) => number;
 export const __wbg_set_jsrequesterror_code: (a: number, b: number) => void;
 export const __wbg_get_jsrequesterror_message: (a: number) => [number, number];
 export const __wbg_set_jsrequesterror_message: (a: number, b: number, c: number) => void;
-export const __wbg_jspasskeyservice_free: (a: number, b: number) => void;
-export const jspasskeyservice_get_credentials_request_options: () => any;
+export const start: () => void;
 export const __wbg_loginservice_free: (a: number, b: number) => void;
 export const __wbg_checkqrcodeloginresponse_free: (a: number, b: number) => void;
 export const __wbg_get_checkqrcodeloginresponse_scan_user_info: (a: number) => number;
@@ -90,6 +90,6 @@ export const __wbindgen_malloc: (a: number, b: number) => number;
 export const __wbindgen_realloc: (a: number, b: number, c: number, d: number) => number;
 export const __wbindgen_export_6: WebAssembly.Table;
 export const __externref_drop_slice: (a: number, b: number) => void;
-export const closure207_externref_shim: (a: number, b: number, c: any) => void;
-export const closure219_externref_shim: (a: number, b: number, c: any, d: any) => void;
+export const closure201_externref_shim: (a: number, b: number, c: any) => void;
+export const closure213_externref_shim: (a: number, b: number, c: any, d: any) => void;
 export const __wbindgen_start: () => void;
